+++ includes/class.llms.l10n.js.php
@@ -7,7 +7,7 @@
  *
  * @package  LifterLMS/Classes/Localization
  * @since    3.17.8
- * @version  7.1.1
+ * @version  7.2.0
  */
 
 defined( 'ABSPATH' ) || exit;
@@ -33,7 +33,7 @@ class LLMS_L10n_JS {
 	 * @param    array $strings existing strings from core / 3rd parties.
 	 * @return   array
 	 * @since    3.17.8
-	 * @version  7.1.1
+	 * @version  7.2.0
 	 */
 	public function get_strings( $strings ) {
 		// phpcs:disable
@@ -299,7 +299,7 @@ class LLMS_L10n_JS {
 			 * File: assets/js/llms-analytics.js.
 			 *
 			 * @since    3.0.0
-			 * @version  3.0.0
+			 * @version  7.2.0
 			 */
 			'Filter by Student(s)' => esc_html__( 'Filter by Student(s)', 'lifterlms' ),
 			'Error' => esc_html__( 'Error', 'lifterlms' ),