+++ templates/admin/reporting/tabs/widgets.php
@@ -5,13 +5,15 @@
  * @package LifterLMS/Templates/Admin
  *
  * @since Unknown
- * @version Unknown
+ * @since 7.2.0 Add content tag param to widget options.
+ * @version 7.2.0
  */
 
 defined( 'ABSPATH' ) || exit;
 if ( ! is_admin() ) {
 	exit;
 }
+
 ?>
 
 <?php foreach ( $widget_data as $row => $widgets ) : ?>
@@ -27,7 +29,14 @@ if ( ! is_admin() ) {
 					<a href="<?php echo esc_url( $opts['link'] ); ?>">
 				<?php } ?>
 
-				<h1><?php echo $opts['content']; ?></h1>
+				<?php
+				printf(
+					'<%s class="llms-widget-content">%s</%s>',
+					esc_html( $opts['content_tag'] ?? 'h3' ),
+					esc_html( $opts['content'] ?? '' ),
+					esc_html( $opts['content_tag'] ?? 'h3' )
+				);
+				?>
 
 				<?php if ( ! empty( $opts['link'] ) ) { ?>
 					</a>